Move flags to be the center of their tiles.
Previously, flags are placed at the top left corner of tiles.
Graphically, flags occupy entire tiles, but behind the scenes,
locations are tracked in floating-point.
This means that glolfers will aim at the top left corner of
flag tiles, and sometimes this causes them to constantly "miss",
oscillating between the tile above and to the left because the
ball doesn't quite enter the flag tile.